refactor: re-organize view files
(+ clean unused code)
