--- polls/classes/poll_option_form.py
@@ -1,10 +1,0 @@
-from polls.models.poll_option_model import PollOptionModel
-from django.forms import ModelForm
-
-class PollForm(ModelForm):
-    class Meta:
-        model = PollOptionModel
-        fields="__all__"
-        labels={
-            "value": "Valore opzione", 
-        }
