+++ polls/templates/polls/majority-results-old.html
@@ -1,0 +1,72 @@
+{% extends 'global/skeleton.html' %}
+
+{% block pagename %}Visualizza i risultati{% endblock pagename %}
+
+{% block content %}
+<div class="container text-center m-5">
+    <h1>{{poll_results.poll_option_data.poll_fk.name}}</h1>
+    <h2 class="text-info">Risultati sondaggio</h2>
+</div>
+
+
+<div class="container">
+
+    
+    <div class="container mt-5 mb-5">
+        <p class="mb-0">Testo della domanda: </p>
+        <!--
+        <h3>{{ poll_results.poll.question }}</h3>
+        -->
+    </div>
+    
+
+    <div class="container mb-3 mt-3">
+        <h4>Risultati</h4>	
+        <div class="container">
+            <h2>Legenda</h2>
+            <span style="background-color:#FF6200;" class="label">1</span>
+            <span style="background-color:#FD7F2C;" class="label">2</span>
+            <span style="background-color:#FD9346;" class="label">3</span>
+            <span style="background-color:#FDA766;" class="label">4</span>
+            <span style="background-color:#FDB777;" class="label">5</span>            
+        </div>
+        <div class="container m-3">
+            <span class="fw-bold">{{poll_results.poll_option_data.value}}</span>
+            <div class="progress">
+                <div style="background-color:#FF6200; width: 50%" class="progress-bar" role="progressbar" aria-valuenow="50" aria-valuemin="0" aria-valuemax="100">50%</div>
+                <div style="background-color:#FD7F2C; width: 30%" class="progress-bar" role="progressbar" aria-valuenow="30" aria-valuemin="0" aria-valuemax="100">30%</div>
+                <div style="background-color:#FD9346; width: 10%" class="progress-bar" role="progressbar" aria-valuenow="10" aria-valuemin="0" aria-valuemax="100">10%</div>
+                <div style="background-color:#FDA766; width: 5%" class="progress-bar" role="progressbar"  aria-valuenow="5" aria-valuemin="0" aria-valuemax="100">5%</div>
+                <div style="background-color:#FDB777; width: 5%" class="progress-bar" role="progressbar"  aria-valuenow="5" aria-valuemin="0" aria-valuemax="100">5%</div>
+            </div>
+        </div>
+        <div class="container m-3">
+            <span class="fw-bold">Opzione 1</span>
+            <div class="progress">
+                <div style="background-color:#FF6200; width: 50%" class="progress-bar" role="progressbar" aria-valuenow="50" aria-valuemin="0" aria-valuemax="100">50%</div>
+                <div style="background-color:#FD7F2C; width: 30%" class="progress-bar" role="progressbar" aria-valuenow="30" aria-valuemin="0" aria-valuemax="100">30%</div>
+                <div style="background-color:#FD9346; width: 10%" class="progress-bar" role="progressbar" aria-valuenow="10" aria-valuemin="0" aria-valuemax="100">10%</div>
+                <div style="background-color:#FDA766; width: 5%" class="progress-bar" role="progressbar"  aria-valuenow="5" aria-valuemin="0" aria-valuemax="100">5%</div>
+                <div style="background-color:#FDB777; width: 5%" class="progress-bar" role="progressbar"  aria-valuenow="5" aria-valuemin="0" aria-valuemax="100">5%</div>
+            </div>
+        </div>
+    </div>
+
+
+    <div class="container text-center m-5">
+        <a name="" id="" role="button"
+            class="btn btn-primary" 
+            href="{% url 'home' %}" >
+            Torna alla home
+        </a>
+        <a name="" id="" role="button"
+            class="btn btn-secondary" 
+            href="https://github.com/SDF-Ingegneria-2022/smoothPoll/issues" 
+            target="_blank">
+            Lasciaci un Feedback
+        </a> 
+    </div>
+</div>
+</div>
+{% endblock content %}
+
