+++ polls/templates/polls/majority-results.html
@@ -3,8 +3,8 @@
 {% block pagename %}Visualizza i risultati{% endblock pagename %}
 
 {% block content %}
-<div class="container text-center m-5">
-    <h1>{{poll_results.poll_option_data.poll_fk.name}}</h1>
+<div class="container text-center mt-5 mb-5">
+    <h1>{{ poll.name }}</h1>
     <h2 class="text-info">Risultati sondaggio</h2>
 </div>
 
@@ -14,41 +14,26 @@
     
     <div class="container mt-5 mb-5">
         <p class="mb-0">Testo della domanda: </p>
-        <!--
-        <h3>{{ poll_results.poll.question }}</h3>
-        -->
+        <h3>{{ poll.question }}</h3>
     </div>
     
 
     <div class="container mb-3 mt-3">
-        <h4>Risultati</h4>	
-        <div class="container">
-            <h2>Legenda</h2>
-            <span style="background-color:#FF6200;" class="label">1</span>
-            <span style="background-color:#FD7F2C;" class="label">2</span>
-            <span style="background-color:#FD9346;" class="label">3</span>
-            <span style="background-color:#FDA766;" class="label">4</span>
-            <span style="background-color:#FDB777;" class="label">5</span>            
-        </div>
+        <h4>Risultati</h4>
         <div class="container m-3">
-            <span class="fw-bold">{{poll_results.poll_option_data.value}}</span>
-            <div class="progress">
-                <div style="background-color:#FF6200; width: 50%" class="progress-bar" role="progressbar" aria-valuenow="50" aria-valuemin="0" aria-valuemax="100">50%</div>
-                <div style="background-color:#FD7F2C; width: 30%" class="progress-bar" role="progressbar" aria-valuenow="30" aria-valuemin="0" aria-valuemax="100">30%</div>
-                <div style="background-color:#FD9346; width: 10%" class="progress-bar" role="progressbar" aria-valuenow="10" aria-valuemin="0" aria-valuemax="100">10%</div>
-                <div style="background-color:#FDA766; width: 5%" class="progress-bar" role="progressbar"  aria-valuenow="5" aria-valuemin="0" aria-valuemax="100">5%</div>
-                <div style="background-color:#FDB777; width: 5%" class="progress-bar" role="progressbar"  aria-valuenow="5" aria-valuemin="0" aria-valuemax="100">5%</div>
-            </div>
-        </div>
-        <div class="container m-3">
-            <span class="fw-bold">Opzione 1</span>
-            <div class="progress">
-                <div style="background-color:#FF6200; width: 50%" class="progress-bar" role="progressbar" aria-valuenow="50" aria-valuemin="0" aria-valuemax="100">50%</div>
-                <div style="background-color:#FD7F2C; width: 30%" class="progress-bar" role="progressbar" aria-valuenow="30" aria-valuemin="0" aria-valuemax="100">30%</div>
-                <div style="background-color:#FD9346; width: 10%" class="progress-bar" role="progressbar" aria-valuenow="10" aria-valuemin="0" aria-valuemax="100">10%</div>
-                <div style="background-color:#FDA766; width: 5%" class="progress-bar" role="progressbar"  aria-valuenow="5" aria-valuemin="0" aria-valuemax="100">5%</div>
-                <div style="background-color:#FDB777; width: 5%" class="progress-bar" role="progressbar"  aria-valuenow="5" aria-valuemin="0" aria-valuemax="100">5%</div>
-            </div>
+            {%for voted_option in poll_results.get_sorted_options %}
+                <div class="row">
+                    <div class="col-6">
+                        <p class="text-secondary">{{forloop.counter}}.
+                            <span class="{% if forloop.counter == 1 %}text-success fw-bold
+                            {% else %}text-dark{% endif %}">{{voted_option.option.value}}</span></p>
+                    </div>
+                    <div class="col-6">
+                        <p class="fw-bold text-primary"> ha ricevuto {{voted_option.n_votes}} voti</p>
+                    </div>
+                </div>
+                <hr />
+            {% endfor %}
         </div>
     </div>
