+++ polls/views/__init__.py
@@ -1,2 +1,3 @@
 from .dummy_view import *
 from .polls_create_view import *
+from .majority_view import *
