+++ polls/views/dummy_view.py
@@ -140,80 +140,6 @@ def results(request: HttpRequest, poll_id: int):
         {'poll_results': poll_results}
         )
 
-def dummy_majority(request: HttpRequest): 
-    """
-    Dummy poll page, here user can try to vote.
-    
-    """
-    options_selected = request.session.get('majvote-submit-error')
-    if options_selected is not None:
-        del request.session['majvote-submit-error']
-
-    try:
-        poll = PollModel.objects.filter(poll_type='majority_vote').first()
-    except Exception:
-        raise Http404
-    return render(
-                    request, 
-                    'polls/majority-vote.html', 
-                    {
-                        'poll': poll, 
-                        
-                        'error': {
-                                     'message': "Attenzione! Non è stata selezionata nessuna opzione.",
-                                     'options_selected': options_selected,
-                                 },
-                        'prova': {'1': {
-                                        '2':2
-                                        }
-                                }
-                        }
-                    )        
-
-def majority_vote_submit(request: HttpRequest, poll_id: int):
-    """Render page with confirmation of majority vote validation."""
-
-    ratings: List[dict] = []
-    session_object: dict = {
-        'id': []
-    }
-    #poll: PollModel = PollModel.objects.filter(poll_id=poll_id)
-    for key, value in request.POST.items():
-        if not key == 'csrfmiddlewaretoken':
-            rating: dict = {}
-            rating["poll_choice_id"] = int(key)
-            rating["rating"] = int(value)
-            ratings.append(rating)
-            session_object['id'].append(int(key))
-            session_object[int(key)] =  int(value)
-
-
-    try:
-        vote: MajorityVoteModel = MajorityVoteService.perform_vote(ratings, poll_id=str(poll_id))
-    except PollOptionRatingUnvalidException:
-       
-        request.session['majvote-submit-error'] = session_object
-        return HttpResponseRedirect(reverse('polls:dummy_majority' ))
-    except Exception as e:
-        raise Http404
-    return render(request, 'polls/vote-majority-confirm.html', {'vote': vote})
-
-def majority_vote_results(request: HttpRequest, poll_id: int):
-    """Render page with majority poll results"""
-
-    try:
-        poll_results: List[MajorityPollResultData] = MajorityVoteService.calculate_result(poll_id=str(poll_id))
-    except PollDoesNotExistException:
-        raise Http404
-    except Exception:
-        # Internal error: you should inizialize DB first (error 500)
-        return HttpResponseServerError("Dummy survey is not initialized. Please see README.md and create it.")
-
-    return render(request, 'polls/majority-results.html', 
-        
-        {'poll_results': poll_results}
-        )
-
 def all_polls(request: HttpRequest):
     """
     Render page with all polls.
